requirements.txt
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
decord==0.6.0
ffmpeg-python==0.2.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
gdown==5.2.0
huggingface-hub==0.29.1
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
narwhals==1.28.0
networkx==3.4.2
numpy==2.2.3
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
portalocker==3.1.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==309
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
# Editable install with no version control (SAM-2==1.0)
-e d:\sam2\sam2
scenedetect==0.6.5.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.1.0
six==1.17.0
soupsieve==2.6
statsmodels==0.14.4
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vidgear==0.3.3
wheel==0.44.0
